Make length BIGINT

+++ radiofeed/episodes/models.py
@@ -81,7 +81,7 @@ class Episode(models.Model):
 
     media_url = models.URLField(max_length=1000)
     media_type = models.CharField(max_length=60)
-    length = models.IntegerField(null=True, blank=True)
+    length = models.BigIntegerField(null=True, blank=True)
 
     duration = models.CharField(max_length=30, blank=True)
     explicit = models.BooleanField(default=False)
